LPD-80236 Move color picker types
